+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "next_boilerplate_styled_pages",
+  "name": "blog_evandro_strapi_frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
