+++ src/templates/AboutTemplate/AboutTemplate.tsx
@@ -1,7 +1,7 @@
 import AboutSection from '../../components/About/AboutSession/AboutSection';
 import { Setting as SettingStrapi } from '../../types/strapi';
 import { About as AboutStrapi } from '../../types/strapi-about';
-import Base from '../BaseTemplate';
+import Base from '../BaseTemplate/BaseTemplate';
 
 export interface AboutTemplateProps {
   setting: SettingStrapi;
