+++ src/templates/index.ts
@@ -1,0 +1,5 @@
+export * from './AboutTemplate/AboutTemplate';
+export * from './AuthorTemplate/AuthorTemplate';
+export * from './BaseTemplate/BaseTemplate';
+export * from './PostTamplate/PostTamplate';
+export * from './PostsTemplate/PostsTemplate';
